Title: 최대 숫자 만들기, Time: 1120ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 240522/최대 숫자 만들기/make-biggest-num.java
@@ -8,8 +8,8 @@ class Type implements Comparable<Type>{
     }
     @Override
     public int compareTo(Type t){
-        double left = Double.parseDouble(Integer.toString(this.num) + Integer.toString(t.num));
-        double right = Double.parseDouble(Integer.toString(t.num) + Integer.toString(this.num));
+        Long left = Long.parseLong(Integer.toString(this.num) + Integer.toString(t.num));
+        Long right = Long.parseLong(Integer.toString(t.num) + Integer.toString(this.num));
         if(left > right){
             return -1;
         }
